moves and removes
