--- model.json
@@ -1,1 +1,0 @@
-{"class_name": "Model", "config": {"name": "model_8", "layers": [{"name": "input", "class_name": "InputLayer", "config": {"batch_input_shape": [null, 32, 2], "dtype": "float32", "sparse": false, "name": "input"}, "inbound_nodes": []}, {"name": "lstm_36", "class_name": "LSTM", "config": {"name": "lstm_36", "trainable": true, "return_sequences": false, "return_state": false, "go_backwards": false, "stateful": false, "unroll": false, "implementation": 0, "units": 128, "activation": "tanh", "recurrent_activation": "hard_sigmoid", "use_bias": true, "kernel_initializer": {"class_name": "VarianceScaling", "config": {"scale": 1.0, "mode": "fan_avg", "distribution": "uniform", "seed": null}}, "recurrent_initializer": {"class_name": "Orthogonal", "config": {"gain": 1.0, "seed": null}}, "bias_initializer": {"class_name": "Zeros", "config": {}}, "unit_forget_bias": true, "kernel_regularizer": null, "recurrent_regularizer": null, "bias_regularizer": null, "activity_regularizer": null, "kernel_constraint": null, "recurrent_constraint": null, "bias_constraint": null, "dropout": 0.0, "recurrent_dropout": 0.0}, "inbound_nodes": [[["input", 0, 0, {}]]]}, {"name": "output_notes", "class_name": "Dense", "config": {"name": "output_notes", "trainable": true, "units": 1, "activation": "softmax", "use_bias": true, "kernel_initializer": {"class_name": "VarianceScaling", "config": {"scale": 1.0, "mode": "fan_avg", "distribution": "uniform", "seed": null}}, "bias_initializer": {"class_name": "Zeros", "config": {}}, "kernel_regularizer": null, "bias_regularizer": null, "activity_regularizer": null, "kernel_constraint": null, "bias_constraint": null}, "inbound_nodes": [[["lstm_36", 0, 0, {}]]]}, {"name": "output_volume", "class_name": "Dense", "config": {"name": "output_volume", "trainable": true, "units": 1, "activation": "sigmoid", "use_bias": true, "kernel_initializer": {"class_name": "VarianceScaling", "config": {"scale": 1.0, "mode": "fan_avg", "distribution": "uniform", "seed": null}}, "bias_initializer": {"class_name": "Zeros", "config": {}}, "kernel_regularizer": null, "bias_regularizer": null, "activity_regularizer": null, "kernel_constraint": null, "bias_constraint": null}, "inbound_nodes": [[["lstm_36", 0, 0, {}]]]}], "input_layers": [["input", 0, 0]], "output_layers": [["output_notes", 0, 0], ["output_volume", 0, 0]]}, "keras_version": "2.0.8", "backend": "tensorflow"}
